% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limix-python.R
\name{install_limix_dependencies}
\alias{install_limix_dependencies}
\title{Install limix}
\usage{
install_limix_dependencies(
  envname = "limix",
  method = "auto",
  flash_whl = FALSE,
  ...
)
}
\arguments{
\item{envname}{virtual environment name}

\item{method}{method defaults to "auto"}

\item{flash_whl}{logical, whether to try installing flash-attn wheel}

\item{...}{additional passes for \code{py_install}}
}
\value{
installs to "limix" env by default
}
\description{
LimiX itself is not a package so this is a very experimental binding.
Install LimiX dependencies will install all python dependencies and clone LimiX.
LimiX should be cloned to the working directory.
}
