% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limix-fit.R
\name{limix_fit}
\alias{limix_fit}
\title{LimiX Fit Interface}
\usage{
limix_fit(
  X_train,
  y_train,
  device = "cuda",
  limix_model = "16M",
  limix_dir = paste0(here::here(), "/LimiX"),
  limix_config = "cls_default_noretrieval"
)
}
\arguments{
\item{X_train}{a set of predictors}

\item{y_train}{a outcome variable}

\item{device}{either 'cuda' or 'cpu'}

\item{limix_model}{this should be "16M" or "2M"}

\item{limix_dir}{based on where you have a clone of LimiX (should be cloned `here::here()``)}

\item{limix_config}{this is a string name of one of the special json of LDM defaults
something like 'cls_default_2M_retrieval', 'reg_default_16M_retrieval', or '{type}\emph{default}{size}_(no)retrieval'}
}
\value{
a limix model fit spec object
}
\description{
LimiX Fit Interface
}
