% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limix-pred.R
\name{limix_predict}
\alias{limix_predict}
\title{LimiX predictions}
\usage{
limix_predict(object, X_test, task_type = "Classification", ...)
}
\arguments{
\item{object}{an object returned from limix_fit}

\item{X_test}{a set of predictors}

\item{task_type}{either \code{Classification} or \code{Regression}}

\item{...}{pass through for prediction method}
}
\value{
predictions
}
\description{
LimiX predictions
}
