% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limix-python.R
\name{use_limix_env}
\alias{use_limix_env}
\title{Use Virtual Environment Wrapper}
\usage{
use_limix_env(envname = "limix", ...)
}
\arguments{
\item{envname}{virtual environment to use}

\item{...}{additional passes for \code{use_virtualenv}}
}
\value{
sets environment to virtualenv
}
\description{
Use Virtual Environment Wrapper
}
